+++ snuba/consumers/consumer.py
@@ -1,5 +1,6 @@
 import itertools
 import logging
+import random
 import time
 from collections import defaultdict
 from datetime import datetime
@@ -32,10 +33,7 @@ from arroyo.processing.strategies import (
     CommitOffsets,
     FilterStep,
     ParallelTransformStep,
-)
-from arroyo.processing.strategies import ProcessingStrategy
-from arroyo.processing.strategies import ProcessingStrategy as ProcessingStep
-from arroyo.processing.strategies import (
+    ProcessingStrategy,
     ProcessingStrategyFactory,
     Reduce,
     RunTaskInThreads,
@@ -57,6 +55,7 @@ from confluent_kafka import Message as ConfluentMessage
 from confluent_kafka import Producer as ConfluentKafkaProducer
 from confluent_kafka import Producer as ConfluentProducer
 
+from snuba import environment, state
 from snuba.clickhouse.http import JSONRow, JSONRowEncoder, ValuesRowEncoder
 from snuba.consumers.schemas import get_json_codec
 from snuba.consumers.types import KafkaMessageMetadata
@@ -75,6 +74,8 @@ from snuba.utils.metrics.wrapper import MetricsWrapper
 from snuba.utils.streams.configuration_builder import build_kafka_producer_configuration
 from snuba.utils.streams.topics import Topic as SnubaTopic
 from snuba.writer import BatchWriter
+
+metrics = MetricsWrapper(environment.metrics, "consumer")
 
 logger = logging.getLogger("snuba.consumer")
 
@@ -103,16 +104,13 @@ class BytesInsertBatch(NamedTuple):
             return type(self), (self.rows, self.origin_timestamp)
 
 
-class InsertBatchWriter(ProcessingStep[BytesInsertBatch]):
+class InsertBatchWriter:
     def __init__(self, writer: BatchWriter[JSONRow], metrics: MetricsBackend) -> None:
         self.__writer = writer
         self.__metrics = metrics
 
         self.__messages: MutableSequence[Message[BytesInsertBatch]] = []
         self.__closed = False
-
-    def poll(self) -> None:
-        pass
 
     def submit(self, message: Message[BytesInsertBatch]) -> None:
         assert not self.__closed
@@ -179,23 +177,17 @@ class InsertBatchWriter(ProcessingStep[BytesInsertBatch]):
             self.__writer,
         )
 
-    def terminate(self) -> None:
-        self.__closed = True
-
     def join(self, timeout: Optional[float] = None) -> None:
         pass
 
 
-class ReplacementBatchWriter(ProcessingStep[ReplacementBatch]):
+class ReplacementBatchWriter:
     def __init__(self, producer: ConfluentKafkaProducer, topic: Topic) -> None:
         self.__producer = producer
         self.__topic = topic
 
         self.__messages: MutableSequence[Message[ReplacementBatch]] = []
         self.__closed = False
-
-    def poll(self) -> None:
-        pass
 
     def submit(self, message: Message[ReplacementBatch]) -> None:
         assert not self.__closed
@@ -226,9 +218,6 @@ class ReplacementBatchWriter(ProcessingStep[ReplacementBatch]):
                     on_delivery=self.__delivery_callback,
                 )
 
-    def terminate(self) -> None:
-        self.__closed = True
-
     def join(self, timeout: Optional[float] = None) -> None:
         args = []
         if timeout is not None:
@@ -249,7 +238,7 @@ class ProcessedMessageBatchWriter:
     def __init__(
         self,
         insert_batch_writer: InsertBatchWriter,
-        replacement_batch_writer: Optional[ProcessingStep[ReplacementBatch]] = None,
+        replacement_batch_writer: Optional[ReplacementBatchWriter] = None,
         # If commit log config is passed, we will produce to the commit log topic
         # upon closing each batch.
         commit_log_config: Optional[CommitLogConfig] = None,
@@ -260,12 +249,6 @@ class ProcessedMessageBatchWriter:
         self.__offsets_to_produce: MutableMapping[Partition, Tuple[int, datetime]] = {}
 
         self.__closed = False
-
-    def poll(self) -> None:
-        self.__insert_batch_writer.poll()
-
-        if self.__replacement_batch_writer is not None:
-            self.__replacement_batch_writer.poll()
 
     def submit(
         self, message: Message[Union[None, BytesInsertBatch, ReplacementBatch]]
@@ -323,14 +306,6 @@ class ProcessedMessageBatchWriter:
                 )
                 self.__commit_log_config.producer.poll(0.0)
         self.__offsets_to_produce.clear()
-
-    def terminate(self) -> None:
-        self.__closed = True
-
-        self.__insert_batch_writer.terminate()
-
-        if self.__replacement_batch_writer is not None:
-            self.__replacement_batch_writer.terminate()
 
     def join(self, timeout: Optional[float] = None) -> None:
         start = time.time()
@@ -428,10 +403,6 @@ class MultistorageCollector:
         ] = defaultdict(list)
         self.__offsets_to_produce: MutableMapping[Partition, Tuple[int, datetime]] = {}
 
-    def poll(self) -> None:
-        for step in self.__steps.values():
-            step.poll()
-
     def submit(
         self,
         message: Message[
@@ -462,12 +433,6 @@ class MultistorageCollector:
 
         for storage_key, step in self.__steps.items():
             step.close()
-
-    def terminate(self) -> None:
-        self.__closed = True
-
-        for step in self.__steps.values():
-            step.terminate()
 
     def join(self, timeout: Optional[float] = None) -> None:
         start = time.time()
@@ -551,13 +516,37 @@ def process_message(
     processor: MessageProcessor,
     consumer_group: str,
     snuba_logical_topic: SnubaTopic,
-    validate: bool,
     message: Message[KafkaPayload],
 ) -> Union[None, BytesInsertBatch, ReplacementBatch]:
+    sentry_sdk.set_tag("snuba_logical_topic", snuba_logical_topic.name)
+
+    validate_sample_rate = float(
+        state.get_config(f"validate_schema_{snuba_logical_topic.name}", 0) or 0.0
+    )
+
     assert isinstance(message.value, BrokerValue)
     try:
         codec = get_json_codec(snuba_logical_topic)
-        decoded = codec.decode(message.payload.value, validate=validate)
+        should_validate = random.random() < validate_sample_rate
+        start = time.time()
+
+        decoded = codec.decode(message.payload.value, validate=False)
+        if should_validate:
+            try:
+                codec.validate(decoded)
+            except Exception as err:
+                sentry_sdk.set_tag("invalid_message_schema", "true")
+                logger.warning(err, exc_info=True)
+
+            # TODO: this is not the most efficient place to emit a metric, but
+            # as long as should_validate is behind a sample rate it should be
+            # OK.
+            metrics.timing(
+                "codec_decode_and_validate",
+                (time.time() - start) * 1000,
+                tags={"snuba_logical_topic": snuba_logical_topic.name},
+            )
+
         result = processor.process_message(
             decoded,
             KafkaMessageMetadata(
@@ -780,7 +769,7 @@ class MultistorageConsumerProcessingStrategyFactory(
 
         transform_function = self.__process_message_fn
 
-        inner_strategy: ProcessingStep[MultistorageKafkaPayload]
+        inner_strategy: ProcessingStrategy[MultistorageKafkaPayload]
 
         if self.__processes is None:
             inner_strategy = TransformStep(transform_function, collect)