+++ snuba/consumers/schemas.py
@@ -1,8 +1,26 @@
+import json
+import os
 from typing import Any, Mapping, MutableMapping, Optional
 
-from arroyo.processing.strategies.decoder import JsonCodec
+from arroyo.processing.strategies.decoder.json import JsonCodec
 
 from snuba.utils.streams.topics import Topic
+
+
+def load_file(filename: str) -> Any:
+    with open(os.path.join(os.path.dirname(__file__), "schema_files/", filename)) as f:
+        return json.load(f)
+
+
+_HARDCODED_SCHEMAS: Mapping[Topic, Mapping[str, Any]] = {
+    # Querylog JSON
+    Topic.QUERYLOG: load_file("querylog.json"),
+    # Release health metrics
+    Topic.METRICS: load_file("metrics.json"),
+    # XXX(markus): This is copypasted from Relay, need to consolidate at some
+    # point
+    Topic.EVENTS: load_file("event.json"),
+}
 
 
 def get_schema(topic: Topic) -> Optional[Mapping[str, Any]]:
@@ -13,7 +31,7 @@ def get_schema(topic: Topic) -> Optional[Mapping[str, Any]]:
     This function returns either the schema if it is defined, or None if not.
 
     """
-    return None
+    return _HARDCODED_SCHEMAS.get(topic)
 
 
 _cache: MutableMapping[Topic, JsonCodec] = {}