+++ tests/datasets/entities/test_pluggable_entity.py
@@ -124,6 +124,8 @@ def build_request(query_body: Mapping[str, str]) -> Request:
         query=query,
         snql_anonymized=snql_anonymized,
         query_settings=HTTPQuerySettings(referrer=""),
-        attribution_info=AttributionInfo(get_app_id("blah"), "blah", None, None, None),
+        attribution_info=AttributionInfo(
+            get_app_id("blah"), {"tenant_type": "tenant_id"}, "blah", None, None, None
+        ),
     )
     return request