+++ tests/datasets/test_querylog_processor.py
@@ -46,7 +46,12 @@ def test_simple() -> None:
         snql_anonymized="",
         query_settings=HTTPQuerySettings(referrer="search"),
         attribution_info=AttributionInfo(
-            get_app_id("default"), "search", None, None, None
+            get_app_id("default"),
+            {"tenant_type": "tenant_id"},
+            "search",
+            None,
+            None,
+            None,
         ),
     )
 
@@ -167,7 +172,12 @@ def test_missing_fields() -> None:
         snql_anonymized="",
         query_settings=HTTPQuerySettings(referrer="search"),
         attribution_info=AttributionInfo(
-            get_app_id("default"), "search", None, None, None
+            get_app_id("default"),
+            {"tenant_type": "tenant_id"},
+            "search",
+            None,
+            None,
+            None,
         ),
     )
 
@@ -289,7 +299,12 @@ def test_negative_project_id_fields() -> None:
         snql_anonymized="",
         query_settings=HTTPQuerySettings(referrer="search"),
         attribution_info=AttributionInfo(
-            get_app_id("default"), "search", None, None, None
+            get_app_id("default"),
+            {"tenant_type": "tenant_id"},
+            "search",
+            None,
+            None,
+            None,
         ),
     )